requirements.txt
psycopg2==2.8.4
Django==1.11.*
Pillow==5.3.0
Unidecode==1.1.1
django-mptt==0.9.0
django-tinymce==2.7.0
django-modeladmin-reorder==0.3.1
django-admin-sortable2==0.7.3
# установить системно postgresql 9